Move android sponsor viewmodel to separate file
